requirements.txt
grpcio==1.54.0
grpcio-tools==1.54.0
protobuf==4.22.3
pytest>=7.0.0
pytest-cov>=4.0.0
